requirements.txt
absl-py==2.3.1
altgraph==0.17.4
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==3.0.0
astunparse==1.6.3
beautifulsoup4==4.13.4
blinker==1.8.2
cachetools==5.3.1
certifi==2025.7.9
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.4.2
clang==17.0.6
click==8.1.7
colorama==0.4.6
comm==0.2.2
comtypes==1.2.0
contourpy==1.3.2
cryptography==42.0.8
cssselect2==0.7.0
customtkinter==5.2.1
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.14
decorator==4.4.2
deepface==0.0.93
distlib==0.3.6
Django==5.1.3
django-cors-headers==4.2.0
django-extensions==3.2.3
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
executing==2.2.0
facenet-pytorch==2.5.3
fer==22.5.1
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.18.0
fire==0.7.0
Flask==3.0.3
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.58.5
fsspec==2025.5.1
future==1.0.0
gast==0.6.0
gdown==5.2.0
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
graphviz==0.20.1
grpcio==1.73.1
gunicorn==23.0.0
h5py==3.14.0
html5lib==1.1
httplib2==0.22.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
imutils==0.5.4
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jinja2-time==0.2.0
joblib==1.4.2
jsonpickle==0.9.6
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
lxml==5.2.2
lz4==4.4.4
make==0.1.6.post2
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
moviepy==1.0.3
mpmath==1.3.0
mtcnn==1.0.0
mysql-connector-python==8.4.0
mysqlclient==2.2.4
mysqlx-connector-python==8.4.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.5
node==1.2.1
nodejs==0.1.1
Nuitka==1.9.7
numpy==2.1.3
oauth2client==4.1.2
odict==1.9.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
optional-django==0.1.0
optree==0.16.0
ordered-set==4.1.0
oscrypto==1.3.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
pefile==2023.2.7
pillow==11.3.0
platformdirs==3.8.0
plumber==1.7
proglog==0.1.12
prompt_toolkit==3.0.51
protobuf==5.29.5
psutil==7.0.0
psycopg==3.2.12
psycopg-binary==3.2.12
pure_eval==0.2.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycparser==2.22
pydot==1.4.2
pydot-ng==2.0.0
pygame==2.5.2
Pygments==2.19.2
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.12
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
pypdf==4.3.0
pypiwin32==223
pypng==0.20220715.0
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyttsx3==2.90
pytz==2025.2
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==27.0.0
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.4
retina-face==0.0.17
rich==14.0.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
six==1.17.0
soupsieve==2.7
SpeechRecognition==3.10.0
sqlparse==0.5.2
stack-data==0.6.3
svglib==1.5.1
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tf_keras==2.19.0
threadpoolctl==3.5.0
tinycss2==1.3.0
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.2
uritemplate==3.0.1
uritools==4.0.3
urllib3==2.5.0
virtualenv==20.23.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
xhtml2pdf==0.2.16
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==6.0
zope.interface==6.4.post2
zope.lifecycleevent==5.0
zope.proxy==5.2
zstandard==0.22.0
